Now i know what is definition and declaration of function
