+++ homework/vector-of-shared-ptrs/vectorFunctions.cpp
@@ -1,15 +1,40 @@
 #include "vectorFunctions.hpp"
-#include <iostream>
-#include <string>
 
+std::vector<std::shared_ptr<int>> generate(int count) {
+    std::vector<std::shared_ptr<int>> generated_vector;
+    generated_vector.reserve(count);
+    for (int value = 0; value < count; value++) {
+        generated_vector.push_back(std::make_shared<int>(value));
+    }
+    return generated_vector;
+}
 
-int main() {
-    auto vec = generate(10);
-    print(vec);
-    add10(vec);
-    print(vec);
-    sub10(vec);
-    print(vec);
+void print(std::vector<std::shared_ptr<int>> vec) {
+    for (auto value : vec) {
+        if (value != nullptr) {
+            std::cout << *value << '\t'; 
+        }
+    }
+}
 
-    return 0;
+void add10(std::vector<std::shared_ptr<int>> vec) {
+    for (auto value : vec) {
+        if (value != nullptr) {
+            *value += 10;
+        }
+    }
+}
+
+void sub10(int *const ptr) {
+    if( ptr != nullptr) {
+        *ptr -= 10;
+    }
+}
+
+void sub10(std::vector<std::shared_ptr<int>> vec) {
+    for(auto value : vec) {
+        if (value != nullptr){
+            sub10(value.get());
+        }
+    }
 }
